shill: Adapt to libchrome TimeDelta factory changes
The libchrome uprev CL:7510045 changed base::TimeDelta's
factory functions (like base::Milliseconds) to use
base::ClampedNumeric.

This causes a compile error when passing an unnamed enum like
dbus::ObjectProxy::TIMEOUT_USE_DEFAULT because ClampedNumeric's
template deduction fails for it.

Explicitly cast it to int to fix the build.

BUG=none
TEST=None

Cq-Depend: chromium:7510045
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7512987
Tested-by: Jeff Lin <jeffulin@google.com>
Commit-Queue: Brett Brotherton <bbrotherton@google.com>

+++ shill/dbus/client/client.h
@@ -24,8 +24,8 @@
 
 namespace shill {
 
-constexpr base::TimeDelta kDefaultDBusTimeout =
-    base::Milliseconds(dbus::ObjectProxy::TIMEOUT_USE_DEFAULT);
+constexpr base::TimeDelta kDefaultDBusTimeout = base::Milliseconds(
+    static_cast<int>(dbus::ObjectProxy::TIMEOUT_USE_DEFAULT));
 
 // Shill D-Bus client for listening to common manager, service and device
 // properties. This class is the result of an effort to consolidate a lot of
